Renaming ppp-crypto.* to crypto.* and pppcrypt.* to crypto_ms.*. This should improve clarity in the crypto subsystem. Also adding comments to the crypto.h as it is a public API
