+++ pppd/mppe.c
@@ -40,7 +40,7 @@
 #include "ccp.h"
 #include "chap_ms.h"
 #include "mppe.h"
-#include "ppp-crypto.h"
+#include "crypto.h"
 
 u_char mppe_send_key[MPPE_MAX_KEY_SIZE];
 u_char mppe_recv_key[MPPE_MAX_KEY_SIZE];
