+++ pppd/ppp-des.c
@@ -43,7 +43,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "ppp-crypto-priv.h"
+#include "crypto-priv.h"
 
 #ifdef OPENSSL_HAVE_DES
